test scene modified, added camera controller, modified invisible walls to stop teleporting the player out of the scene, added energy consumption

+++ DarkLight/Assets/Scripts/PlayerEnergyController.cs
@@ -17,7 +17,7 @@ public class PlayerEnergyController : MonoBehaviour {
         energyLevel = initialEnergyLevel;
 	}
 
-    public void AddEnergy(int quantity)
+    public void AddEnergy(float quantity)
     {
         if (energyLevel < initialEnergyLevel)
         {
